cambiosviernes31

+++ app/Livewire/AbsenceComponent.php
@@ -15,46 +15,73 @@ class AbsenceComponent extends Component
     public $absence;
     public $description;
     public $absenceDate;
+    public $department;
     public $hour;
     public $turn;
     public $user_id;
     public $absenceForm = false;
     public $absence_id;
-    public $teacher_absences=[];
+    public $teacher_absences = [];
     public $name;
-    
-    public function mount(){
-        $this->absences=$this->getAbsences();
+    public $departments = [];
+    public $hours = [];
+    public $turns = [];
+    public $date;
+    public function mount()
+    {
+
+        $this->departments = $this->getDepartments();
     }
     public function render()
     {
 
-      //  $this->absences = DB::table('absences')->get();
- 
-      $this->absences = DB::table('absences')->join('users','absences.user_id','=','users.id')
-      ->join('departments','departments.id','=','users.department_id')->select('absences.*','users.*','departments.*')->get();
+        //  $this->absences = DB::table('absences')->get();
+
+        $this->absences = DB::table('absences')->join('users', 'absences.user_id', '=', 'users.id')
+            ->join('departments', 'departments.id', '=', 'users.department_id')->select('absences.*', 'users.*', 'departments.*')->get();
         return view('livewire.absence-component');
+    }
+    public function getAbsencesDepsTeachers()
+    {
+
+        $this->absences = DB::table('absences')->join('users', 'absences.user_id', '=', 'users.id')
+            ->join('departments', 'departments.id', '=', 'users.department_id')->select('absences.*', 'users.*', 'departments.*')->get();
     }
     public function createAbsence()
     {
+        $id=User::where('name','==',$this->name);
         $absence = new Absence();
         $absence->description = $this->description;
         $absence->hour = $this->hour;
         $absence->turn = $this->turn;
-        $absence->absenceDate=$this->absenceDate;
+        $absence->absenceDate = $this->absenceDate;
         $absence->user_id = auth()->id();
         $absence->save();
-        $this->getAbsences();
-        /*
-'description',
-        'hour',
-        'turn',
-        'user_id',
-
-*/
+        $this->getAbsencesDepsTeachers();
+        $this->closeAbsenceForm();
     }
-  
+    public function getDepartments()
+    {
+        $this->departments = DB::table('departments')->select('dep_name')->get();
+    }
+    public function clearFields()
+    {
+        $this->absenceDate = '';
+        $this->department = '';
+        $this->hour = '';
+        $this->turn = '';
+        $this->name = '';
+        $this->description = '';
+    }
 
+
+    /*
+public $description;
+    public $absenceDate;
+    public $hour;
+    public $turn;
+    name
+*/
     public function updateAbsence()
     {
         $user = auth()->id();
@@ -65,7 +92,7 @@ class AbsenceComponent extends Component
             'description' => $this->description,
             'hour' => $this->hour,
             'turn' => $this->turn,
-            'absenceDate'=>$this->absenceDate,
+            'absenceDate' => $this->absenceDate,
             'user_id' => $user,
         ]);
         $this->displayTeachersAbsences();
@@ -74,13 +101,13 @@ class AbsenceComponent extends Component
     }
     public function displayTeachersAbsences()
     {
-        $this->teacher_absences = Absence::with('user')->where('user_id','=',auth()->id())->get();
+        $this->teacher_absences = Absence::with('user')->where('user_id', '=', auth()->id())->get();
     }
 
-  
+
     public function openAbsenceForm()
     {
-
+        $this->getDepartments();
         $this->absenceForm = true;
     }
     public function closeAbsenceForm()
@@ -101,12 +128,15 @@ class AbsenceComponent extends Component
     }
 
 
-    public function deleteAbsence(Absence $absence)
+    public function deleteAbsence($absenceId)
     {
-        $absence->delete();
-        $this->displayTeachersAbsences();
+        $absence = Absence::find($absenceId);
+        if ($absence) {
+            $absence->delete();
+            $this->getAbsencesDepsTeachers();
+        }
     }
-   /* public function getAbsences2()
+    /* public function getAbsences2()
     {
         $this->absences = DB::table('absences')->get();
     }*/
